// Vesta/frontend/src/app/shared/models/buildingPerson.model.ts
export class BuildingPerson {
  id?: string;
  objectId?: string;
  displayName: string;
  street: number;
  number: number;
  zipcode: number;
  city: string;
  email: string;
  phone: string;
  additionalInfo: string;
  role: string;
  authorizationType?: string;
  authStartDate?: Date;
  authEndDate?: Date;
}